Add annotations to MARK
We want some kind of meta level to group and reason about entities and rules.
One use case we currently solve with comments is the tracking of rules for
specific standards and guidelines. This is somewhat cumbersome.

We add a meta language layer comparable to Java's annotation feature.
Annotated entities and rules could now be queried by their annotations.
This allows to reason about the set of all rules with a common annotation.
Additionally, this would allow to define new evaluation rules based on
annotations. E.g. some generic requirement is fulfilled if some or all
rules with a given annotation are evaluated positively.