Make MARK file content more optional (#9)
We usually require that a MARK file contains at least an entity or a rule. However,
there are cases where we comment out rules because they are very use case dependent.
This causes parsing errors in MARK.

We change the minimal structure of MARK files to now require only a `package`
declaration.